+++ FlexColorPicker/Classes/Utilities/LimitedGestureViewDelegate.swift
@@ -26,6 +26,6 @@
 //  SOFTWARE.
 //
 
-public protocol LimitedGestureViewDelegate {
+public protocol LimitedGestureViewDelegate: class {
     func gestureRecognizerShouldBegin(_ gestureRecognizer: UIGestureRecognizer) -> Bool
 }